Rename again for consistency with ResponseLoggingMiddleware
